+++ app/[domain]/(blank)/upgrade-to-pro/page.tsx
@@ -1,3 +1,6 @@
+import { DEFAULT_UPGRADE_TO_PRO_SECTION } from '$lib/components/features/upgrade-to-pro/sections';
+import UpgradeToProPage from './UpgradeToProPage';
+
 import UpgradeToProContainer from '$lib/components/features/upgrade-to-pro/UpgradeToProContainer';
 import { GetListSubscriptionDocument, SubscriptionItem } from '$lib/graphql/generated/backend/graphql';
 import { getClient } from '$lib/graphql/request';
@@ -7,5 +10,7 @@ export default async function Page() {
   const { data } = await client.query({ query: GetListSubscriptionDocument });
   const subscriptionData = (data?.listSubscriptionItems || []) as SubscriptionItem[];
 
-  return <UpgradeToProContainer subscriptionData={subscriptionData} />;
+  return <UpgradeToProPage activeSection={DEFAULT_UPGRADE_TO_PRO_SECTION} />;
+
+  // return <UpgradeToProContainer subscriptionData={subscriptionData} activeSection={DEFAULT_UPGRADE_TO_PRO_SECTION} />;
 }